Properly compile functions.
Before we would pull them out and compile them one by one. Now we compile then
as we compile the entire AST, but save the compilation result until the AST is
finished, and then we gather them all. This is important because arrow
functions need to retain their correct scopes.