Rewrite scopes to be more generic about how we enter/leave scopes.
Basically we want to do scope resolution by type symbol, when possible, because
we sometimes want to have scopes available even when we don't have the actual
nodes. For example, if you do an import of a class, you want all the methods in
scope, even though you don't have the class node.